[ refactor ] useToast 훅 위치 변경
