+++ src/@components/CardCollectionPage/index.tsx
@@ -8,9 +8,9 @@ import HeaderMinVer from "../@common/Header/HeaderMinVer";
 import useGTMPage from "../@common/hooks/useGTMPage";
 import useModal from "../@common/hooks/useModal";
 import useScroll from "../@common/hooks/useScrollToTop";
-import useToast from "../@common/hooks/useToast";
 import LoginModal from "../@common/LoginModal";
 import SuspenseBoundary from "../@common/SuspenseBoundary";
+import useToast from "../@common/Toast/hooks/useToast";
 import Toast from "../@common/Toast/ToastProvider";
 import MenuModal from "./Card/MenuModal";
 import CardSlider from "./CardSlider";
